Import of mport 2.6.8

+++ libmport/audit.c
@@ -1,5 +1,5 @@
 /*-
- * SPDX-License-Identifier: BSD-2-Clause-FreeBSD
+ * SPDX-License-Identifier: BSD-2-Clause
  *
  * Copyright (c) 2023 Lucas Holt
  * All rights reserved.
@@ -43,7 +43,8 @@ MPORT_PUBLIC_API char *
 mport_audit(mportInstance *mport, const char *packageName, bool dependOn)
 {
 	mportPackageMeta **packs = NULL;
-	mportPackageMeta **depends, **depends_orig = NULL;
+	mportPackageMeta **depends = NULL;
+	mportPackageMeta **depends_orig = NULL;
 	char *pkgAudit = NULL;
 	struct ucl_parser *parser = NULL;
 
@@ -58,6 +59,9 @@ mport_audit(mportInstance *mport, const char *packageName, bool dependOn)
 	}
 
 	if (mport_pkgmeta_search_master(mport, &packs, "pkg=%Q", packageName) != MPORT_OK) {
+		mport_pkgmeta_vec_free(packs);
+		packs = NULL;
+
 		return (NULL);
 	}
 
@@ -67,8 +71,14 @@ mport_audit(mportInstance *mport, const char *packageName, bool dependOn)
 			char *jsonData = readJsonFile(path);
 			if (jsonData == NULL) {
 				SET_ERROR(MPORT_ERR_FATAL, "Error opening CVE file");
-				unlink(path);
-				free(path);
+				if (path != NULL) {
+					unlink(path);
+					free(path);
+					path = NULL;
+				}
+				mport_pkgmeta_vec_free(packs);
+				packs = NULL;
+
 				return (NULL);
 			}
 
@@ -82,8 +92,14 @@ mport_audit(mportInstance *mport, const char *packageName, bool dependOn)
 				ucl_parser_free(parser);
 
 				free(jsonData);
+				jsonData = NULL;
+
 				unlink(path);
 				free(path);
+				path = NULL;
+
+				mport_pkgmeta_vec_free(packs);
+				packs = NULL;
 
 				return (NULL);
 			}
@@ -99,8 +115,15 @@ mport_audit(mportInstance *mport, const char *packageName, bool dependOn)
 			if (bufferFp == NULL) {
 				SET_ERROR(MPORT_ERR_FATAL, "Error allocating memory for audit entries");
 				free(jsonData);
+				jsonData = NULL;
+
 				unlink(path);
 				free(path);
+				path = NULL;
+				
+				mport_pkgmeta_vec_free(packs);
+				packs = NULL;
+
 				return (NULL);
 			}
 
@@ -108,6 +131,33 @@ mport_audit(mportInstance *mport, const char *packageName, bool dependOn)
 			while ((cur = ucl_object_iterate(root, &it, true))) {
 				if (ucl_object_type(cur) != UCL_OBJECT) {
 					SET_ERROR(MPORT_ERR_FATAL, "Expected an object in the array");
+					continue;
+				}
+
+				bool isVulnerable = false;
+				const ucl_object_t *products = ucl_object_find_key(cur, "products");
+				if (products != NULL && ucl_object_type(products) == UCL_ARRAY) {
+					ucl_object_iter_t pit = NULL;
+					const ucl_object_t *product;
+
+					for (product = ucl_object_iterate(products, &pit, true); product != NULL; product = ucl_object_iterate(products, &pit, true)) {
+						const ucl_object_t *version = ucl_object_find_key(product, "version");
+						if (version != NULL && ucl_object_type(version) == UCL_STRING) {
+							const char *version_str = ucl_object_tostring(version);
+
+							if (version_str == NULL)
+							    continue;
+
+							// Skip based on the version field
+							if ('*' == version_str[0] || mport_version_cmp((*packs)->version, version_str) <= 0) {
+								isVulnerable = true;
+								break;
+							}
+						}
+					}
+				}
+
+				if (!isVulnerable) {
 					continue;
 				}
 
@@ -150,14 +200,22 @@ mport_audit(mportInstance *mport, const char *packageName, bool dependOn)
 						fprintf(bufferFp, "\n");
 
 						mport_pkgmeta_vec_free(depends_orig);
+						depends_orig = NULL;
+						depends = NULL;
 					}
 				}
 			}
 
 			free(jsonData);
+			jsonData = NULL;
+
 			fclose(bufferFp);
+			bufferFp = NULL;
+
 			unlink(path);
 			free(path);
+			path = NULL;
+
 			ucl_object_unref(root);
 
 			mport_pkgmeta_vec_free(packs);
